+++ testsuite/six-dir/using_six.py
@@ -1,4 +1,4 @@
 from six.moves import html_parser
 from six.moves import reload_module
-from six.moves import reduce
-from six.moves.urllib.error import URLError
+from functools import reduce
+from urllib.error import URLError