+++ testsuite/six-dir/using_six_incorrectly.py
@@ -1,1 +1,1 @@
-from six.moves import copyreg
+import copyreg