+++ doc/changelog.rst
@@ -4,6 +4,9 @@ Release history
 2.3
 ---
 
+* Supports python 3.10 or later (drops support for 3.9
+  and adds support for 3.13 and 3.14)
+
 * Add `modulegraph2.Modulegraph.add_dependencies_for_source``
   that will add the imports in a python code fragment
   to a graph as roots.
@@ -29,6 +32,12 @@ Release history
 
   This currently only supports packages found in the
   filesystem (excluding zipfiles).
+
+* Type annotation for the argument of ``ModuleGraph.add_excludes``
+  is now ``Iterable[str]`` instead of ``Iterator[str]``.
+
+* Added ``extension_attributes`` to ``PyPIDistribution`` for
+  consistency with other nodes.
 
 2.2.1
 -----